requirements.txt
absl-py
Pillow
pudb
torch
matplotlib
torchvision
PyYAML
expecttest
easydict
tensorboard
opencv-python
pudb